+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - Improvements to class generation and imports to compendiums. (Such as primary abilities and being set at level 20).
+- When importing an Adventure to a compendium, depending on setting selections, it could delete the folders in the world.
 
 # 6.3.7